SC-18650: keep video job module ratchet green
